android-patch-48 (#57)
* android: improve code

* android: fix #48
