+++ dtb-android/app/config.gradle
@@ -1,7 +1,7 @@
 ext.appName = 'DTBazar'
 ext.versionMajor = 1
 ext.versionMinor = 35
-ext.versionPatch = 43
+ext.versionPatch = 48
 ext.versionSuffix = ''
 ext.vars = [
         apiHost          : 'API_HOST',
