+++ dtb-android/app/src/main/java/app/roaim/dtbazar/ui/home/HomeViewModel.kt
@@ -15,12 +15,15 @@ class HomeViewModel @Inject constructor(
 ) : ViewModel() {
 
     private val _uid = MutableLiveData<String>()
-
     private val _myStores = MediatorLiveData<List<Store>>()
     private val _myDonations = MediatorLiveData<List<Donation>>()
+    private val _logout = MutableLiveData<Boolean>()
+
+    val logout: LiveData<Boolean> = _logout
 
     val profile: LiveData<Result<Profile>> = infoRepository.getProfile().map {
         if (it.status == Status.SUCCESS) _uid.postValue(it.data?.id)
+        if (it.status == Status.LOGOUT) _logout.postValue(true)
         it
     }
 
@@ -32,14 +35,12 @@ class HomeViewModel @Inject constructor(
         }
         val myRemoteDonations = donationRepository.getMyDonations()
         _myDonations.addSource(myRemoteDonations) {
-            if (it.status == Status.SUCCESS) {
-                _myDonations.value = it.data
-                _myDonations.removeSource(myRemoteDonations)
-            }
+            if (it.status == Status.SUCCESS) _myDonations.value = it.data
+            if (it.status == Status.LOGOUT) _logout.postValue(true)
+            if (it.status != Status.LOADING) _myDonations.removeSource(myRemoteDonations)
         }
         _myDonations
     }
-
 
     val myStores = _uid.switchMap { uid ->
         if (uid != null) _myStores.addSource(storeRepository.getMyCachedStores(uid)) {
@@ -47,34 +48,29 @@ class HomeViewModel @Inject constructor(
         }
         val myRemoteStores = storeRepository.getMyStores()
         _myStores.addSource(myRemoteStores) {
-            if (it.status == Status.SUCCESS) {
-                _myStores.value = it.data
-                _myStores.removeSource(myRemoteStores)
-            }
+            if (it.status == Status.SUCCESS) _myStores.value = it.data
+            if (it.status == Status.LOGOUT) _logout.postValue(true)
+            if (it.status != Status.LOADING) _myStores.removeSource(myRemoteStores)
         }
         _myStores
     }
-
 
     fun saveStore(
         name: String,
         address: String,
         mobile: String,
         target: LatLng?
-    ): LiveData<Result<Store>> =
-            storeRepository.saveStore(
-                StorePostBody(
-                    name = name,
-                    address = address,
-                    mobile = mobile,
-                    location = listOf(
-                        target?.latitude ?: ipInfo.value?.lat,
-                        target?.longitude ?: ipInfo.value?.lon
-                    )
-                )
+    ): LiveData<Result<Store>> = storeRepository.saveStore(
+        StorePostBody(
+            name = name,
+            address = address,
+            mobile = mobile,
+            location = listOf(
+                target?.latitude ?: ipInfo.value?.lat,
+                target?.longitude ?: ipInfo.value?.lon
             )
-
+        )
+    )
 
     fun deleteStore(store: Store): LiveData<Result<Store>> = storeRepository.deleteStore(store)
-
 }
